+++ packages/geolocation/test/server.test.ts
@@ -1,16 +1,34 @@
-import { createGeolocation, createGeolocationWatcher } from "../src/index.js";
+import {
+  makeGeolocation,
+  makeGeolocationWatcher,
+  createGeolocation,
+  createGeolocationWatcher,
+} from "../src/index.js";
 import { describe, expect, it } from "vitest";
 
 describe("API doesn't break in SSR", () => {
+  it("makeGeolocation() - SSR", async () => {
+    const [query, cleanup] = makeGeolocation();
+    expect(cleanup).toBeInstanceOf(Function);
+    await expect(query()).rejects.toThrow();
+  });
+
+  it("makeGeolocationWatcher() - SSR", () => {
+    const [store, cleanup] = makeGeolocationWatcher();
+    expect(store.location).toBe(null);
+    expect(store.error).toBe(null);
+    expect(cleanup).toBeInstanceOf(Function);
+  });
+
   it("createGeolocation() - SSR", () => {
     const [location, refetch] = createGeolocation();
-    expect(location()).toBe(undefined);
-    expect(location.loading).toBe(true);
-    expect(location.error).toBe(undefined);
+    expect(() => location()).toThrow();
     expect(refetch).toBeInstanceOf(Function);
   });
 
   it("createGeolocationWatcher() - SSR", () => {
-    expect(createGeolocationWatcher(true)).toEqual({ location: null, error: null });
+    const { location, error } = createGeolocationWatcher(true);
+    expect(() => location()).toThrow();
+    expect(error()).toBe(null);
   });
 });